+++ src/game/tests.rs
@@ -411,6 +411,7 @@ mod landwalk;
 mod mana_and_costs;
 mod mana_vault;
 mod modal_effects;
+mod morbid;
 mod old_school_interactions;
 mod old_school_permanents;
 mod old_school_spells;